refactor: #41 폴더 구조 변경 (styles/tokens 폴더 추가)
